requirements.txt
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arize==7.36.0
astroid==3.3.9
attrs==25.1.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.46.0
Deprecated==1.2.18
dill==0.3.9
docker==7.1.0
fastapi==0.115.11
Flask==3.1.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mlflow==2.21.0
mlflow-skinny==2.21.0
mypy-extensions==1.0.0
narwhals==1.30.0
numpy==1.26.4
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pylint==3.3.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-futures==1.0.0
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.1
streamlit==1.43.1
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.0
zipp==3.21.0
